# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DogMap API",
    "description": "API REST para sistema de locais pet-friendly.\n\n## Sobre o Projeto\nO DogMap é uma plataforma que permite aos usuários descobrir, avaliar e favoritar locais que aceitam pets.\n\n## Funcionalidades Principais\n- **Usuários**: Cadastro e gerenciamento de usuários com informações de seus pets\n- **Locais**: Catálogo de estabelecimentos pet-friendly com diferentes categorias\n- **Avaliações**: Sistema de avaliações e comentários para os locais\n- **Favoritos**: Lista personalizada de locais favoritos por usuário\n\n## Tecnologias\n- Node.js + Express\n- Dados mockados para desenvolvimento\n- Swagger para documentação",
    "version": "1.0.0",
    "contact": {
      "name": "DogMap Team",
      "email": "contato@dogmap.com",
      "url": "https://dogmap.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3009",
      "description": "Servidor de Desenvolvimento"
    },
    {
      "url": "https://api.dogmap.com",
      "description": "Servidor de Produção"
    }
  ],
...
  "paths": {
...
    "/api/users": {
      "get": {
        "summary": "Lista todos os usuários",
        "description": "Retorna uma lista completa de todos os usuários cadastrados no sistema",
        "tags": ["Usuários"],
        "responses": {
          "200": {
            "description": "Lista de usuários retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    },
                    "count": {
                      "type": "integer",
                      "description": "Número total de usuários",
                      "example": 5
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "summary": "Cadastra um novo usuário",
        "description": "Cria um novo usuário com informações pessoais e dados do pet",
        "tags": ["Usuários"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "pet"],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Nome completo do usuário",
                    "example": "João Silva",
                    "minLength": 2,
                    "maxLength": 100
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email único do usuário",
                    "example": "joao.silva@email.com"
                  },
                  "pet": {
                    "$ref": "#/components/schemas/PetInput"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Usuário criado com sucesso"
                    },
                    "data": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "summary": "Busca um usuário por ID",
        "description": "Retorna as informações detalhadas de um usuário específico",
        "tags": ["Usuários"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID único do usuário",
            "schema": {
              "type": "string"
            },
            "example": "user_123"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Pet": {
        "type": "object",
        "description": "Informações sobre o pet do usuário",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome do pet",
            "example": "Rex"
          },
          "type": {
            "type": "string",
            "description": "Tipo do pet",
            "example": "Cachorro"
          },
          "breed": {
            "type": "string",
            "description": "Raça do pet",
            "example": "Golden Retriever"
          }
        },
        "required": ["name", "type", "breed"]
      },
      "PetInput": {
        "type": "object",
        "description": "Dados para cadastro de um pet",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nome do pet",
            "example": "Rex",
            "minLength": 1,
            "maxLength": 50
          },
          "type": {
            "type": "string",
            "description": "Tipo do pet",
            "example": "Cachorro",
            "enum": ["Cachorro", "Gato", "Pássaro", "Peixe", "Outros"]
          },
          "breed": {
            "type": "string",
            "description": "Raça do pet",
            "example": "Golden Retriever",
            "minLength": 1,
            "maxLength": 50
          }
        },
        "required": ["name", "type", "breed"]
      },
      "User": {
        "type": "object",
        "description": "Informações do usuário",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único do usuário",
            "example": "user_123"
          },
          "name": {
            "type": "string",
            "description": "Nome completo do usuário",
            "example": "João Silva"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email do usuário",
            "example": "joao.silva@email.com"
          },
          "pet": {
            "$ref": "#/components/schemas/Pet"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data de criação do usuário",
            "example": "2024-01-15T10:30:00Z"
          }
        },
        "required": ["id", "name", "email", "pet"]
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Dados inválidos ou obrigatórios não fornecidos",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "error": {
                  "type": "string",
                  "example": "Dados obrigatórios não fornecidos"
                },
                "message": {
                  "type": "string",
                  "example": "Nome, email e informações do pet são obrigatórios"
                }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Recurso não encontrado",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "error": {
                  "type": "string",
                  "example": "Usuário não encontrado"
                },
                "message": {
                  "type": "string",
                  "example": "Usuário com ID user_123 não foi encontrado"
                }
              }
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Erro interno do servidor",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "error": {
                  "type": "string",
                  "example": "Erro interno do servidor"
                },
                "message": {
                  "type": "string",
                  "example": "Ocorreu um erro inesperado"
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT para autenticação (implementação futura)"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}